DB 연동 및 오류 수정
캘린더 리스트 불러올 때 오류 수정 및 세부페이지 데이터 연동

+++ app/src/main/java/com/example/bookreport/data/model/Record.kt
@@ -8,7 +8,7 @@ data class Record (
     @SerializedName("image") val image: String,
     @SerializedName("author") val author: String,
     @SerializedName("publisher") val publisher: String,
-    @SerializedName("rating") val rating: Double,
+    @SerializedName("rating") val rating: Float,
     @SerializedName("memo") val memo: String,
     @SerializedName("date") val date: String
 )
